+++ Code/Libraries/Foundation/Replication/Peer.cpp
@@ -119,7 +119,7 @@ bool Peer::operator<(Guid rhs) const
 
 ProtocolId Peer::GetProtocolId()
 {
-  static const ProtocolId protocolId(String("Replication Peer").Hash());
+  static const ProtocolId protocolId((uint)String("Replication Peer").Hash());
   return protocolId;
 }
